+++ pypowerwall/tedapi/ted_api_messages/__init__.py
@@ -1,0 +1,16 @@
+import json
+import os
+from .. import tedapi_pb2
+from .ConfigMessage import ConfigMessage
+from .DeviceControllerMessage import DeviceControllerMessage
+from .FirmwareMessage import FirmwareMessage
+from .GatewayStatusMessage import GatewayStatusMessage
+from .TEDAPIMessage import TEDAPIMessage
+
+__all__ = [
+        "ConfigMessage",
+        "DeviceControllerMessage",
+        "FirmwareMessage",
+        "GatewayStatusMessage", 
+        "TEDAPIMessage"
+    ]
