+++ pypowerwall/tedapi/ted_api_messages/graphql/device_controller_query.sig
