+++ pypowerwall/tedapi/ted_api_messages/graphql/status_query.graphql
